requirements.txt
streamlit~=1.34.0
requests~=2.32.2
langchain~=0.2.0
PyMySQL~=1.1.1
SQLAlchemy~=2.0.30
beautifulsoup4~=4.12.3
chroma-hnswlib~=0.7.3
chromadb~=0.5.0
fastapi~=0.111.0
fastapi-cli~=0.0.4
huggingface-hub~=0.23.1
langchain~=0.2.0
langchain-chroma~=0.1.1
langchain-community~=0.2.0
langchain-core~=0.2.1
langchain-experimental~=0.0.59
langchainhub~=0.1.15
lxml~=5.2.2
lxml_html_clean~=0.1.1
newspaper3k~=0.2.8
python-dotenv~=1.0.1
requests~=2.32.2
sentence-transformers~=2.7.0
streamlit~=1.34.0
tiktoken~=0.7.0
transformers~=4.41.0